server/orc/requirements.txt
fastapi
uvicorn
pytesseract
pillow
pyautogui
keyboard 
requests
pywin32
python-multipart